* moved logic of file execution to separate class
