+++ backend/Squirrel.SqlService/Squirrel.SqlService.BLL/Services/SqlFormatter/PgSqlParser.py
