+++ backend/Squirrel.SqlService/Squirrel.SqlService.BLL/Squirrel.SqlService.BLL.csproj
@@ -21,7 +21,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="Services\PgSqlParser.py">
+    <None Update="Services\SqlFormatter\PgSqlParser.py">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
